Enhance buffer performance (#69)
* Enhance the buffer, make everything more modular and easier to maintain
