+++ src/threadfinBuffer.go
@@ -10,7 +10,16 @@ import (
 	"time"
 )
 
-func StartThreadfinBuffer(stream *Stream) error {
+type ThreadfinBuffer struct{
+	StreamBuffer
+
+}
+
+func (sb *ThreadfinBuffer) StartBuffer(stream *Stream) error {
+	if err := sb.StreamBuffer.StartBuffer(stream); err != nil {
+		return err
+	}
+
 	stopChan := make(chan struct{})
 	ShowInfo(fmt.Sprintf("Streaming:Buffer:%s", "Threadfin"))
 	ShowInfo("Streaming URL:" + stream.URL)
@@ -64,7 +73,7 @@ func StartThreadfinBuffer(stream *Stream) error {
 			}
 		}
 	}()
-	stream.Buffer.StopChan = stopChan
+	sb.StopChan = stopChan
 	return nil
 }
 
@@ -90,4 +99,40 @@ func selectStreamFromMaster(resp io.ReadCloser) (string, string, error) {
 		return "", "", err
 	}
 	return videoURL, audioURL, nil
+}
+
+func (sb *ThreadfinBuffer) HandleByteOutput(stdOut io.ReadCloser) {
+    sb.StreamBuffer.HandleByteOutput(stdOut)
+}
+
+func (sb *ThreadfinBuffer) PrepareBufferFolder(folder string) error {
+    return sb.StreamBuffer.PrepareBufferFolder(folder)
+}
+
+func (sb *ThreadfinBuffer) GetBufTmpFiles() []string {
+    return sb.StreamBuffer.GetBufTmpFiles()
+}
+
+func (sb *ThreadfinBuffer) GetBufferedSize() int {
+    return sb.StreamBuffer.GetBufferedSize()
+}
+
+func (sb *ThreadfinBuffer) addBufferedFilesToPipe() {
+    sb.StreamBuffer.addBufferedFilesToPipe()
+}
+
+func (sb *ThreadfinBuffer) DeleteOldestSegment() {
+    sb.StreamBuffer.DeleteOldestSegment()
+}
+
+func (sb *ThreadfinBuffer) CheckBufferFolder() (bool, error) {
+    return sb.StreamBuffer.CheckBufferFolder()
+}
+
+func (sb *ThreadfinBuffer) CheckBufferedFile(file string) (bool, error) {
+    return sb.StreamBuffer.CheckBufferedFile(file)
+}
+
+func (sb *ThreadfinBuffer) writeToPipe(file string) error {
+    return sb.StreamBuffer.writeToPipe(file)
 }
